Restructure data preprocessing
